Fix `UnicodeEncodeError` during csv file writing
For some reasons (maybe because I have a fr_FR.UTF-8 install) I was not able to
run this plugin without running into this error:
```
Traceback (most recent call last):
  File "com_github_bennymeg_JLC-Plugin-for-KiCad/thread.py", line 183, in run
    csv_writer.writerow(component.values())
UnicodeEncodeError: 'ascii' codec can't encode character '\xb5' in position 13: ordinal not in range(128)
```

A small fix is to open the file with UTF-8 encoding ; which solves this issue.

I added the same encoding fix to all 3 text files creation just to guard against
this issue coming in the future.