Add limit of parts in BOM rows (#95)
* Add limit of parts in BOM rows

The number of components that are merged into the same row in the BOM file is limited. If there are too many you'll get an "Internal Server Error" when JLCPCB is processing it. 

Their support told me 20 parts per row is the limit but I've tested with 50 and it works, so I split the difference and set it to 30 here. I guess the 20 was only a suggestion :)

* Use variable instead of magic number for BOM-merge limit

* Use configure value for setting limit of merged parts in BOM