Update the code that matches entries in the DB to match documented behavior.
Modify _get_rotation_from_db to match the database as described in the
documentation.

Modify generate_tables to call both _get_rotation_from_db AND
_get_rotation_offset_from_footprint.  It's not clear from the
docs the intended behavior, this patch applies BOTH.  This allows
the footprint to "override" the database.

It may make more sense to see if the footprint has a value,
if it does use it directly, and if not look up in the database.
Either way, it should be clearly documented in the docs which
behavior happens.